PM: Base abstract TicketBroker
TicketBroker is now an abstract contract that is meant to be the base
contract for specific implementations. The first implementation is the
ETHTicketBroker which denominates ticket payments in ETH. An additional
example implementation provided is ERC20TicketBroker which denominates
ticket payments in a ERC20 token.

Tickets now include a generic auxData field replacing the previous
creationTimestamp field. auxData is meant to store auxilary data that is
relevant for a particular TicketBroker implementation. At the moment,
ETHTicketBroker uses auxData to store a creation timestamp.