PM: Integration test skeleton for ticket flow
LivepeerETHTicketBroker is now hooked up with the rest of the Livepeer
contracts. In this commit, the contract updates the recipient's fee pool
for the current round instead of the round during which the ticket was
created which can be done in a separate update. The ticket flow
integration test only performs basic checks and represents the
foundation for additional test cases. Also added a unit test
for burning ETH in the Minter.