[delta-audit] bonding: Various fixes from audit findings (#622)
* Add remappings file for vs code

* bonding: Avoid double-checkpoint on self-bond/unbond

* bonding: Fix solidity version for IBondingVotes

* bonding: Add lastClaimRound to delegator change event

* bonding,treasury: Fix a few documentations

QA Report https://github.com/code-423n4/2023-08-livepeer-findings/issues/150

* bonding: Improve docs about unused _endRound

Reported on: https://github.com/code-423n4/2023-08-livepeer-findings/issues/205

* bonding: Address PR comments

* bonding: Add deprecation notice to slashTranscoder

We've had some many wardens spending their time auditing this
function and the side-effects of it, when it is disabled and
so doesn't really matter. Adding the comment so we don't need
to remember on the next audit to make this disclaimer.